autoRequire as iife script
